working on replacing container for tile map
want to replace the map, tileIndexingMap, with a custom container thats mroe suitable. the issue being a Flood Fill algorithm would take up to 9x longer with a map compared to a vector.

+++ src/Scenes/LevelCreationScene.h
@@ -56,7 +56,7 @@ namespace EWE {
 		double currentMouseXPos;
 		double currentMouseYPos;
 
-		static void createLevel(uint32_t width, uint32_t height);
+		static void createLevel(uint16_t width, uint16_t height);
 
 		void fitToScreen();
